[Refactor] Videos.tsx
