+++ src/components/pages/videos/Videos.tsx
@@ -1,9 +1,10 @@
-import React from "react";
-import { styled } from "@mui/material/styles";
 import Container from "@mui/material/Container";
 import Grid from "@mui/material/Grid";
+import { styled } from "@mui/material/styles";
 import Typography from "@mui/material/Typography";
+import React from "react";
 import YouTube, { YouTubeEvent } from "react-youtube";
+import { videosData } from "./videos.constants";
 
 const PREFIX = "Videos";
 
@@ -33,64 +34,6 @@ const Root = styled("div")(({ theme }) => ({
     margin: theme.spacing(2),
   },
 }));
-
-const videosData = [
-  {
-    title: "Trails Of Culture",
-    iframes: [
-      {
-        id: "s2T9ilEFrWo",
-      },
-      {
-        id: "8N1C7eIF6Hc",
-      },
-      {
-        id: "SQ3eFSClVzg",
-      },
-    ],
-  },
-  {
-    title: "春晓",
-    iframes: [
-      {
-        id: "85sP940mbYQ",
-      },
-      {
-        id: "tysr8x5aFGI",
-      },
-    ],
-  },
-  {
-    title: "春晓 活动日",
-    iframes: [
-      {
-        id: "Aa3MRyYy5b8",
-      },
-      {
-        id: "y0KLTwdnNTY",
-      },
-    ],
-  },
-  {
-    title: "JPP",
-    iframes: [
-      {
-        id: "SamrBf6bS4Y",
-      },
-    ],
-  },
-  {
-    title: "毕业刊",
-    iframes: [
-      {
-        id: "hKVBjBExhrU",
-      },
-      {
-        id: "6xp3VrT80Aw",
-      },
-    ],
-  },
-];
 
 function Videos() {
   const onReady = (event: YouTubeEvent<any>) => {
