+++ src/routes/AppRoutes.tsx
@@ -1,18 +1,18 @@
-import React from "react";
-import { Route, Routes } from "react-router-dom";
-import Home from "components/pages/Home";
-import Login from "components/pages/auth/Login";
-import Verify from "components/pages/auth/Verify";
 import Create from "components/pages/auth/Create";
+import Login from "components/pages/auth/Login";
 import Reset from "components/pages/auth/Reset";
-import PageNotFound from "components/pages/PageNotFound";
-import Graduates from "components/pages/Graduates";
+import Verify from "components/pages/auth/Verify";
 import GraduateDetails from "components/pages/GraduateDetails";
+import Graduates from "components/pages/Graduates";
+import Home from "components/pages/Home";
 import Lecturers from "components/pages/Lecturers";
-import Videos from "components/pages/Videos";
+import Jpp from "components/pages/magazine/chap1/Jpp";
 import ChapIntro from "components/pages/magazine/ChapIntro";
 import Magazine from "components/pages/magazine/Magazine";
-import Jpp from "components/pages/magazine/chap1/Jpp";
+import PageNotFound from "components/pages/PageNotFound";
+import Videos from "components/pages/videos/Videos";
+import React from "react";
+import { Route, Routes } from "react-router-dom";
 
 const AppRoutes: React.FC = () => {
   return (
